UHA_Net/requirements.txt
matplotlib==3.0.3
numpy==1.18.5
pandas==0.24.2
scipy==1.4.1
#torch==1.3.0
torchfile==0.1.0
torchvision==0.4.1
tqdm==4.51.0
SimpleITK==2.0.2
